+++ test/inspec/aem_base_spec.rb
@@ -1,4 +1,6 @@
-require 'spec_helper'
+require './spec_helper'
+
+init_conf
 
 describe package('nokogiri') do
   it { should be_installed.by('gem').with_version('1.6.8.1') }