+++ test/inspec/httpd_spec.rb
@@ -1,4 +1,6 @@
-require 'spec_helper'
+require './spec_helper'
+
+init_conf
 
 apache_package = @hiera.lookup('apache::apache_name', nil, @scope)
 apache_package ||= 'httpd'
@@ -9,15 +11,5 @@ end
 
 describe service('httpd') do
   it { should be_enabled }
-#  it { should be_running }
-end
-
-# TODO: check disapatcher installation.
-# dispatcher module installation differs depending on os and apache version.
-if os[:family] == 'redhat'
-
-
-
-
-
+  #it { should be_running }
 end