+++ test/inspec/publish_spec.rb
@@ -1,4 +1,6 @@
-require 'spec_helper'
+require './spec_helper'
+
+init_conf
 
 aem_base = @hiera.lookup('publish::aem_base', nil, @scope)
 aem_base ||= '/opt'
@@ -9,7 +11,7 @@ aem_port ||= '4503'
 describe file("#{aem_base}/aem") do
   it { should be_directory }
   it { should exist }
-  it { should be_mode 775 }
+  its('mode') { should cmp '00775' }
   it { should be_owned_by 'root' }
   it { should be_grouped_into 'root' }
 end
@@ -17,7 +19,7 @@ end
 describe file("#{aem_base}/aem/publish") do
   it { should be_directory }
   it { should exist }
-  it { should be_mode 775 }
+  its('mode') { should cmp '00775' }
   it { should be_owned_by 'aem-publish' }
   it { should be_grouped_into 'aem-publish' }
 end
@@ -25,7 +27,7 @@ end
 describe file("#{aem_base}/aem/publish/license.properties") do
   it { should be_file }
   it { should exist }
-  it { should be_mode 440 }
+  its('mode') { should cmp '00440' }
   it { should be_owned_by 'aem-publish' }
   it { should be_grouped_into 'aem-publish' }
 end
@@ -33,25 +35,20 @@ end
 describe file("#{aem_base}/aem/publish/aem-publish-#{aem_port}.jar") do
   it { should be_file }
   it { should exist }
-  it { should be_mode 775 }
+  its('mode') { should cmp '00775' }
   it { should be_owned_by 'aem-publish' }
   it { should be_grouped_into 'aem-publish' }
 end
 
-# Service will be renamed to 'aem' on next puppet-aem release.
-# https://github.com/bstopp/puppet-aem/commit/a28d87fbf6bafc81ff00dec1759d8848708f32af
 describe service('aem-publish') do
   it { should_not be_enabled }
-# serverspec is using ps aux | grep -w to determine if a service is running
-# # aem is just a java process, which fails the test
-# it { should be_running }
-
+  it { should_not be_running }
 end
 
 describe file('/etc/puppetlabs/puppet/publish.yaml') do
   it { should be_file }
   it { should exist }
-  it { should be_mode 644 }
+  its('mode') { should cmp '00644' }
   it { should be_owned_by 'root' }
   it { should be_grouped_into 'root' }
 end